Added users database
Need to get login/registration working before i start working on the rest of the pages, i.e my first iteration of the menu page is concidered done. The only functionality remaining is searching and favourites, and since searching is going to implement favourites then that needs to be done which in turn requires user functionality.