Category toggling sort of working
Well they do toggle a state variable containing true/false for each button, with only a small caveat. When the buttons are NOT toggled, i.e background color is white, it counts as true. One would expect the opposite when working with the code even though it probably makes no difference on the UI